Server+client testing, GH actions, take 1
